requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
PyNaCl>=1.4.0
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
pyinstaller>=5.10.0
aiohappyeyeballs>=2.4.4
aiohttp>=3.11.11
aiosignal>=1.3.2
altgraph>=0.17.4
attrs>=24.3.0
cffi>=1.17.1
frozenlist>=1.5.0
future>=1.0.0
idna>=3.10
multidict>=6.1.0
packaging>=24.2
pefile>=2023.2.7
propcache>=0.2.1
pycparser>=2.22
pyinstaller-hooks-contrib>=2024.11
pywin32-ctypes>=0.2.3
setuptools>=75.8.0
flask>=2.0.0
requests>=2.28.0